move grid placement below layout
